lxd/storage/drivers: de-duplicate Mount/UnmountVolume between Pure/PowerFlex
Introduce a new file remote_storage_helpers.go to consolidate all shared
helpers across different remote storage drivers which use Connectors (NVMe,
iSCSI, etc).
